requirements.txt
boto3==1.4.4
botocore==1.5.33
defusedxml==0.5.0
Django==1.10.6
django-rest-auth==0.9.1
django-storages==1.5.2
djangorestframework==3.6.2
docutils==0.13.1
jmespath==0.9.2
oauthlib==2.0.2
olefile==0.44
piexif==1.0.12
Pillow==4.0.0
psycopg2==2.7.1
PyJWT==1.4.2
python-dateutil==2.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
